[PHI CAPI] Add support for registering a new operator, PART1
